+++ src/test/java/com/ecommerce/api/controller/domain/order/OrderEntityControllerIntegrationTest.java
@@ -1,6 +1,6 @@
 package com.ecommerce.api.controller.domain.order;
 
-import com.ecommerce.DatabaseCleanUp;
+import com.ecommerce.config.DatabaseCleanUp;
 import com.ecommerce.domain.order.orderitem.OrderItemWrite;
 import com.ecommerce.domain.product.ProductWrite;
 import com.ecommerce.domain.user.UserWrite;
