+++ src/test/java/com/ecommerce/config/DatabaseCleanUp.java
@@ -1,4 +1,4 @@
-package com.ecommerce;
+package com.ecommerce.config;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
