+++ src/test/java/com/ecommerce/domain/user/service/UserServiceTest.java
@@ -1,6 +1,6 @@
 package com.ecommerce.domain.user.service;
 
-import com.ecommerce.DatabaseCleanUp;
+import com.ecommerce.config.DatabaseCleanUp;
 import com.ecommerce.domain.coupon.CouponWrite;
 import com.ecommerce.domain.coupon.DiscountType;
 import com.ecommerce.domain.user.UserWrite;
